Practice using descendant combinator
This practice using descendant combinator was interesting because it was
mentioned in the lesson that using two type combinators would not work.
Such as divp, what it failed to mention was that you could in fact do
"div p".
So because of this initial confusion, I ended up writing it as
".container .text", which is fine, but it would have been nice to know
that "div p" would work too.
Additionally, if I just wrote "div" the line between the last two p
elements was yellow also, which seemed strange to me.